Restore compatibility with upstream
This merge request embeds a fix to a minor breaking change introduced in neo4j java driver. The project modifier gets instructed to remove a dependency lately introduced and add the two working ones.

The latest development version of neo4j java driver upstream upgraded the reactor dependency to version Dysprosium-SR7. The reactor library vendor also made some modifications in terms of OSGi metadata configuration. An error was identified in terms of avoiding a package from importing. The following merge request was provided to fix this problem. Ones a new release of the reactor library is available, the reactor version used in neo4j java driver has to be increased to that version.