Track fix OSGi 4.0 branch
Among the in-development versions neo4j java driver 4.0 and 4.1, a rather significant change makes the driver and the current OGM version incompatible. Incompatibility is traced to integration tests, but some changes involve a check of client and server-side bolt protocol versions. For now, this repository tracks the revision 4.0 and deploys artifacts from the build.